+++ view/src/app/App.tsx
@@ -1,17 +1,44 @@
-import { Routes, Route } from 'react-router-dom';
+import { Routes, Route, useNavigate } from 'react-router-dom';
 import Home from '../pages/Home';
 import MyPage from '../pages/MyPage';
 import SignIn from '../pages/SignIn';
 import SignUp from '../pages/SignUp';
+import { MutationCache, QueryCache, QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { useAuthActions } from '../shared/store/auth/authActions';
 
 const App = () => {
+  const navigate = useNavigate();
+  const { logout } = useAuthActions();
+  const queryClient = new QueryClient({
+    queryCache: new QueryCache({
+      onError: (error) => {
+        if (error instanceof Error) {
+          if (error.message === '403') {
+            logout();
+            navigate('signin');
+          }
+        }
+      },
+    }),
+    mutationCache: new MutationCache({
+      onError: (error) => {
+        if (error instanceof Error) {
+          if (error.message === '403') {
+            navigate('signin'); // 403 에러 처리
+          }
+        }
+      },
+    }),
+  });
   return (
-    <Routes>
-      <Route path="/" element={<Home />} />
-      <Route path="/mypage" element={<MyPage />} />
-      <Route path="/signin" element={<SignIn />} />
-      <Route path="/signup" element={<SignUp />} />
-    </Routes>
+    <QueryClientProvider client={queryClient}>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/mypage" element={<MyPage />} />
+        <Route path="/signin" element={<SignIn />} />
+        <Route path="/signup" element={<SignUp />} />
+      </Routes>
+    </QueryClientProvider>
   );
 };
