+++ view/src/entities/MyAssetInfo/model/useDeposit.ts
@@ -1,30 +1,20 @@
 import { useMutation } from '@tanstack/react-query';
-import depositApi from '../api/depositApi';
-import { useNavigate } from 'react-router-dom';
-import { useAuthActions } from '../../../shared/store/auth/authActions';
+import withdrawApi from '../api/withdrawApi';
 
-const useDeposit = () => {
-  const navigate = useNavigate();
-  const { logout } = useAuthActions();
-
+const useWithdraw = () => {
   return useMutation({
-    mutationFn: depositApi,
+    mutationFn: withdrawApi,
     onSuccess: () => {
-      alert('입금 성공');
+      alert('출금 성공');
     },
     onError: (error: unknown) => {
       if (error instanceof Error) {
-        if (error.message === '403') {
-          logout();
-          navigate('/signin');
-        } else {
-          alert('입금에 실패했습니다. 다시 시도해주세요.');
-        }
+        alert('출금에 실패했습니다. 다시 시도해주세요.');
       } else {
-        alert('입금에 실패했습니다. 다시 시도해주세요.');
+        alert('출금에 실패했습니다. 다시 시도해주세요.');
       }
     },
   });
 };
 
-export default useDeposit;
+export default useWithdraw;
